+++ src/matrix/inverse_matrix_utils.c
@@ -6,7 +6,7 @@
 /*   By: rpoder <rpoder@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/21 17:50:14 by rpoder            #+#    #+#             */
-/*   Updated: 2022/10/21 18:05:58 by rpoder           ###   ########.fr       */
+/*   Updated: 2022/10/27 13:13:45 by rpoder           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,18 +14,18 @@
 
 
 
-float	ft_calculate_m2_determinant(t_matrix2 m2)
+double	ft_calculate_m2_determinant(t_matrix2 m2)
 {
 	return ((m2.matrix[0][0] * m2.matrix[1][1])
 		- (m2.matrix[0][1] * m2.matrix[1][0]));
 }
 
 
-float	ft_calculate_m3_determinant(t_matrix3 m3)
+double	ft_calculate_m3_determinant(t_matrix3 m3)
 {
 	int		i;
-	float	cofactor;
-	float	sum;
+	double	cofactor;
+	double	sum;
 
 	i = 0;
 	sum = 0;
@@ -38,11 +38,11 @@ float	ft_calculate_m3_determinant(t_matrix3 m3)
 	return (sum);
 }
 
-float	ft_calculate_m4_determinant(t_matrix4 m4)
+double	ft_calculate_m4_determinant(t_matrix4 m4)
 {
 	int			i;
-	float		cofactor;
-	float		sum;
+	double		cofactor;
+	double		sum;
 
 	i = 0;
 	sum = 0;
@@ -55,11 +55,11 @@ float	ft_calculate_m4_determinant(t_matrix4 m4)
 	return (sum);
 }
 
-float	ft_calculate_m3_cofactor(t_matrix3 m3, int i, int j)
+double	ft_calculate_m3_cofactor(t_matrix3 m3, int i, int j)
 {
-	float	cofactor;
+	double	cofactor;
 	t_matrix2	sub_m2;
-	float		minor;
+	double		minor;
 
 	sub_m2 = ft_sub_m3tom2(m3, j, i);
 	minor = ft_calculate_m2_determinant(sub_m2);
@@ -70,11 +70,11 @@ float	ft_calculate_m3_cofactor(t_matrix3 m3, int i, int j)
 	return (cofactor);
 }
 
-float	ft_calculate_m4_cofactor(t_matrix4 m4, int i, int j)
+double	ft_calculate_m4_cofactor(t_matrix4 m4, int i, int j)
 {
-	float		cofactor;
+	double		cofactor;
 	t_matrix3	sub_m3;
-	float		minor;
+	double		minor;
 
 	sub_m3 = ft_sub_m4tom3(m4, j, i);
 	minor = ft_calculate_m3_determinant(sub_m3);
