+++ src/tuple/color_operations.c
@@ -6,13 +6,13 @@
 /*   By: rpoder <rpoder@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/19 12:18:55 by rpoder            #+#    #+#             */
-/*   Updated: 2022/10/19 16:04:56 by rpoder           ###   ########.fr       */
+/*   Updated: 2022/10/27 13:13:45 by rpoder           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_color	ft_create_color(float red, float green, float blue)
+t_color	ft_create_color(double red, double green, double blue)
 {
 	t_color	new_color;
 
@@ -42,7 +42,7 @@ t_color	ft_sub_colors(t_color a, t_color b)
 	return (new_color);
 }
 
-t_color	ft_scale_color(t_color a, float scale)
+t_color	ft_scale_color(t_color a, double scale)
 {
 	t_color	new_color;
